Lab4Done

+++ Labs1sem/3Laba/HomeTask/Hometask.c
@@ -8,12 +8,12 @@ int main() {
 	system("color F0");
 	setlocale(LC_ALL, "RUS");
 	printf("  ¬ариант 18, поиск гипотенузы по сторонам треугольника\n\n");
-	int a, b;
+	float a, b;
 	double result;
 	printf("  ¬ведите первую сторону треугольника:\n\t");
-	scanf_s("%d", &a);
+	scanf_s("%f", &a);
 	printf("  ¬ведите вторую сторону треугольника:\n\t");
-	scanf_s("%d", &b);
+	scanf_s("%f", &b);
 	result = sqrt((a * a) + (b * b));
 	printf("  √ипотенуза равна: %g", result);
 	puts("");
